requirements.txt
langchain
streamlit
pypdf
openai
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_openai
langchain_experimental
ipykernel
langchain-community
-e .